LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/anchor-position-inline-003-expected.txt

spacer

a1‮2‭z

FAIL .target 1 assert_equals:
<div class="target" style="left: anchor(--a1 left)" data-offset-x="10"></div>
offsetLeft expected 10 but got 0
FAIL .target 2 assert_equals:
<div class="target" style="right: anchor(--a1 right)" data-offset-x="40"></div>
offsetLeft expected 40 but got 100
FAIL .target 3 assert_equals:
<div class="target" style="top: anchor(--a1 top)" data-offset-y="20"></div>
offsetTop expected 20 but got 0
PASS .target 4
